# frontend/src/generated/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Starlib API",
    "description": "Backend API for Starlib music management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/metadata/files/{file_path}/artwork": {
      "get": {
        "tags": [
          "metadata"
        ],
        "summary": "Get File Artwork",
        "description": "Get artwork image for an audio file.\n\nArtwork is extracted once and cached to ``<cache_dir>/artwork/``.  Fast\npath returns the cached file without re-reading the audio file.\n\nParameters\n----------\nfile_path : str\n    Relative or absolute path to audio file\nroot_folder : Path\n    Root music folder (injected)\n\nReturns\n-------\nFileResponse\n    Artwork image (JPEG or PNG)\n\nRaises\n------\nHTTPException\n    If file doesn't exist or has no artwork",
        "operationId": "get_file_artwork_api_metadata_files__file_path__artwork_get",
        "parameters": [
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "metadata"
        ],
        "summary": "Update File Artwork",
        "description": "Update artwork for an audio file by uploading an image.\n\nParameters\n----------\nfile_path : str\n    Relative or absolute path to audio file\nfile : UploadFile\n    Uploaded artwork image file\nroot_folder : Path\n    Root music folder (injected)\n\nReturns\n-------\nOperationResponse\n    Operation result\n\nRaises\n------\nHTTPException\n    If file doesn't exist or upload fails",
        "operationId": "update_file_artwork_api_metadata_files__file_path__artwork_post",
        "parameters": [
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_file_artwork_api_metadata_files__file_path__artwork_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "metadata"
        ],
        "summary": "Delete File Artwork",
        "description": "Remove artwork from an audio file.\n\nParameters\n----------\nfile_path : str\n    Relative or absolute path to audio file\nroot_folder : Path\n    Root music folder (injected)\n\nReturns\n-------\nOperationResponse\n    Operation result\n\nRaises\n------\nHTTPException\n    If file doesn't exist or removal fails",
        "operationId": "delete_file_artwork_api_metadata_files__file_path__artwork_delete",
        "parameters": [
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_update_file_artwork_api_metadata_files__file_path__artwork_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_update_file_artwork_api_metadata_files__file_path__artwork_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "new_file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New File Path"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "OperationResponse",
        "description": "Generic operation success/failure response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}